LPS-55106 Create an unified MessageDigestCacheKeyGenerator
