+++ portal-impl/src/META-INF/util-spring.xml
@@ -88,7 +88,9 @@
 					<bean class="com.liferay.portal.cache.key.HashCodeCacheKeyGenerator" />
 				</entry>
 				<entry key="com.liferay.portlet.PortletPreferencesFactoryImpl">
-					<bean class="com.liferay.portal.cache.key.JavaSHA1CacheKeyGenerator" />
+					<bean class="com.liferay.portal.cache.key.MessageDigestCacheKeyGenerator">
+						<constructor-arg value="SHA-1" />
+					</bean>
 				</entry>
 			</map>
 		</property>
